% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artistas}
\alias{artistas}
\title{Nomes e Links do Portfolio Relativo de Musicas de Artistas do Genero
Gospel no Portal de Letras Vagalume.}
\format{
Um conjunto de dados com 348 artistas e 2 variaveis:
\describe{
\item{nome}{nome do artista}
\item{href}{link relativo para o portfolio de musicas do artista}
}
}
\source{
\url{https://www.vagalume.com.br/browse/style/gospel.html}
}
\usage{
artistas
}
\description{
Names and Relative Links of the  Music Portfolio of Artists of the Genre
Gospel on the Vagalume Lyrics Portal.
}
\keyword{datasets}
