% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter-letras.R
\name{dados_atualizados}
\alias{dados_atualizados}
\title{dados_atualizados}
\usage{
dados_atualizados(tabela)
}
\arguments{
\item{tabela}{tipo da tabela: artistas, musicas ou letras.}
}
\value{
Um dataframe atualizado com conteudo de acordo com a tabela
escolhida.
}
\description{
Funcao para leitura de tabelas atualizadas
}
\examples{
dados_atualizados(tabela = 'artistas')
}
