% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{letras}
\alias{letras}
\title{Letras e Nomes de Musicas de todos os Artistas do Genero Gospel no
Portal de Letras Vagalume.}
\format{
Um conjunto de dados com 42487 letras e 3 variaveis:
\describe{
\item{nome}{nome do artista}
\item{musica}{nome da musica}
\item{lyric}{letra da musica}
}
}
\source{
\url{https://www.vagalume.com.br/browse/style/gospel.html}
}
\usage{
letras
}
\description{
Lyrics and Song Names of all Genero Gospel Artists on Portal of
Vagalume Lyrics.
}
\keyword{datasets}
