% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{musicas}
\alias{musicas}
\title{Nomes do Artista e Musicas, com Links Relativos das Letras do Genero
Gospel no Portal de Letras Vagalume.}
\format{
Um conjunto de dados com 42489 musicas e 3 variaveis:
\describe{
\item{nome}{nome do artista}
\item{musica}{nome da musica}
\item{href}{link relativo para a letra da musica}
}
}
\source{
\url{https://www.vagalume.com.br/browse/style/gospel.html}
}
\usage{
musicas
}
\description{
Artist Names and Songs, with Relative Links of Gospel Lyrics Genre
on the Vagalume Lyrics Portal.
}
\keyword{datasets}
